+++ src/views/ProfileDirectory.vue
@@ -66,6 +66,7 @@ async function loadNextPage() {
   display: grid;
   gap: $block-outer-padding;
   grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+  margin-bottom: $block-outer-padding;
 }
 
 .profile-list-item {