+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Hide payment description if recipient is remote account.
 - Prevent sidebar from obscuring header.
 - Fix router link on profile page.
+- Fix incorrect remote subscription status.
 
 ## [2.1.0] - 2023-10-10