+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Convert "Copy link to post" button to a link.
 - Unlock next page button on home timeline in case of error.
 - Remove extra whitespace in posts without text.
+- Remove extra whitespace around "Publish" button.
 
 ### Fixed