+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Use single API call to identify payment sender.
+- Migrate build system to Vite 3.
 
 ### Fixed