+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Display "Mute author" menu item if post is reposted.
+- Fix position of "Show more profiles" button on profile directory page.
 
 ## [2.2.0] - 2023-10-22